split agent and forward into separate files
